// src/helper/enums/ClientLink.ts
enum ClientLink{

    EmailChange='/handyman/email/change',
    FindHandyman = '/client/handyman/find',
    JobsPosted ='/client/jobs/posted',
    ChnageMail = "/client/email/change",
    ChnagePassword = "/client/password/change",
    OfferAccepted = "/client/offers/accepted",
    JobsDeleted = "/client/jobs/deleted", 
    Messages= "/client/messages",
}


export default ClientLink;